+++ ollama/ollama.go
@@ -116,7 +116,7 @@ type EmbeddingsRequest struct {
 // EmbeddingsResponse represents the response structure for the /api/embeddings endpoint.
 type EmbeddingResponse struct {
 	Model           string      `json:"model"`
-	Embeddings      [][]float64 `json:"embeddings"`
+	Embeddings      [][]float32 `json:"embeddings"`
 	TotalDuration   int64       `json:"total_duration,omitempty"`
 	LoadDuration    int64       `json:"load_duration,omitempty"`
 	PromptEvalCount int         `json:"prompt_eval_count,omitempty"`
